cleanup ToolItems
