+++ src/main/java/gregtech/common/items/tool/GrassPathBehavior.java
@@ -37,7 +37,7 @@ public class GrassPathBehavior implements IToolBehavior {
 
         Set<BlockPos> blocks;
         // only attempt to till if the center block is tillable
-        if (world.isAirBlock(pos.up()) && isBlockPathConvertable(stack, world, player, pos)) {
+        if (world.isAirBlock(pos.up()) && isBlockPathConvertible(stack, world, player, pos)) {
             if (aoeDefinition == AoESymmetrical.none()) {
                 blocks = ImmutableSet.of(pos);
             } else {
@@ -50,7 +50,7 @@ public class GrassPathBehavior implements IToolBehavior {
                 if (rayTraceResult.typeOfHit != RayTraceResult.Type.BLOCK) return EnumActionResult.PASS;
                 if (rayTraceResult.sideHit == null) return EnumActionResult.PASS;
 
-                blocks = getPathConvertableBlocks(stack, aoeDefinition, world, player, rayTraceResult);
+                blocks = getPathConvertibleBlocks(stack, aoeDefinition, world, player, rayTraceResult);
                 blocks.add(rayTraceResult.getBlockPos());
             }
         } else return EnumActionResult.PASS;
@@ -71,11 +71,11 @@ public class GrassPathBehavior implements IToolBehavior {
         return EnumActionResult.PASS;
     }
 
-    public static Set<BlockPos> getPathConvertableBlocks(ItemStack stack, AoESymmetrical aoeDefinition, World world, EntityPlayer player, RayTraceResult rayTraceResult) {
-        return ToolHelper.iterateAoE(stack, aoeDefinition, world, player, rayTraceResult, GrassPathBehavior::isBlockPathConvertable);
+    public static Set<BlockPos> getPathConvertibleBlocks(ItemStack stack, AoESymmetrical aoeDefinition, World world, EntityPlayer player, RayTraceResult rayTraceResult) {
+        return ToolHelper.iterateAoE(stack, aoeDefinition, world, player, rayTraceResult, GrassPathBehavior::isBlockPathConvertible);
     }
 
-    private static boolean isBlockPathConvertable(ItemStack stack, World world, EntityPlayer player, BlockPos pos) {
+    private static boolean isBlockPathConvertible(ItemStack stack, World world, EntityPlayer player, BlockPos pos) {
         if (world.isAirBlock(pos.up())) {
             Block block = world.getBlockState(pos).getBlock();
             return block == Blocks.GRASS || block == Blocks.DIRT; // native dirt to path
