+++ src/main/java/gregtech/common/items/tool/HarvestIceBehavior.java
@@ -7,11 +7,15 @@ import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.world.World;
+import net.minecraftforge.event.world.BlockEvent;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
+/**
+ * @see gregtech.common.ToolEventHandlers#onHarvestDrops(BlockEvent.HarvestDropsEvent)
+ */
 public class HarvestIceBehavior implements IToolBehavior {
 
     // ice harvesting is handled in an event elsewhere
